src/third_party/blink/web_tests/fast/animation/scroll-animations/scrolltimeline-creation.html

<!DOCTYPE html>
<style>
.scroller {
  height: 100px;
  width: 100px;
  overflow: scroll;
}

.content {
  height: 500px;
  width: 500px;
}
</style>

<div class='scroller'>
  <div class='content'></div>
</div>

<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>
<script>
// TODO(smcgruer): Remove once 'auto' timeRange is supported.
test(function() {
  const scroller = document.querySelector('.scroller');
  const constructorFunc = function() { new ScrollTimeline(
      { scrollSource: scroller, orientation: 'block' }) };
  assert_throws('NotSupportedError', constructorFunc);
}, 'If the timeRange is unspecified, object construction should fail');
</script>
